AFDP-537 - Services: Admin Workflow Configuration Page -fixed file download and added integration test for that -applied code cleanup and refactoring -improved checking existing files with MD5 digest -adjusted unit and integration tests -fixed unit test, to get hashes from files on startup -bug in unit test resolved

+++ acm-tool-integrations/acm-activiti-configuration/src/test/java/com/armedia/acm/activiti/services/AcmBpmnServiceTest.java
@@ -57,12 +57,13 @@ public class AcmBpmnServiceTest extends EasyMockSupport {
     @BeforeClass
     public static void initialCleanUp() {
         String userHome = System.getProperty("user.home");
-        for (File f : new File(userHome + "/.acm/activiti/versions").listFiles()) {
-            if (f.isFile() && f.getName().startsWith("Test"))
-                f.delete();
+        File versionsFolder = new File(userHome + "/.acm/activiti/versions");
+        if(versionsFolder.exists()) {
+            for (File f : versionsFolder.listFiles()) {
+                if (f.isFile() && f.getName().startsWith("Test"))
+                    f.delete();
+            }
         }
-
-
     }
 
     @Before
